pyproject.toml
[project]
name = "loginfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.2",
    "openai>=2.30.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
